Merge GCal and WebCal into one ICal, add lombok
Lots of misc backend features and fixes as well.

Unfortunately, Aspen is down ATM so I can't work on the new Halfday / FLEX stuff, so it was removed. This is a breaking change, but I'm not changing the json schema version as I hope to have it back as it was before by the time school starts.