Use a public backup of the calendar
The Heroku instance for some reason is unable to connect to the calendar. Try using the WaybackMachine as a proxy. This URL will need to be updated periodically for days to continue to work.

We should find a better fix for this, but it will have to do for now