Clean code and update to be less assumptive
Rewrote a large portion of the code. We now use a ScheduleDocumentParser class to build the schedule, which is more flexible w/ Spring MVC.

Days are no longer assumed to be numbers, we don't assume there are 6 days in a cycle, and Blocks are no longer assumed to be one character.

The Day object stores most of the data, but we still are backwards compatible with the old API via Schedule class. Users with strict typing may need to update their clients, but the AspenAppWebframe for MHS should be unaffected, as it uses plain JS.